REFERRAL: Rename views & blocks + small SCSS changes to padding
