+++ src/components/blocks/HeroReferee/HeroReferee.scss
@@ -1,6 +1,6 @@
 @import '../../../styles/main';
 
-.invite {
+.referee {
   @include background-image;
   @include col;
   @include padding(top, 6);
