+++ src/components/blocks/index.js
@@ -4,9 +4,9 @@ export { default as Featured } from './Featured/Featured';
 export { default as Footer } from './Footer/Footer';
 export { default as Header } from './Header/Header';
 export { default as Hero } from './Hero/Hero';
-export { default as HeroInvite } from './HeroInvite/HeroInvite';
+export { default as HeroReferee } from './HeroReferee/HeroReferee';
 export { default as Image } from './Image/Image';
 export { default as ProductShowcase } from './ProductShowcase/ProductShowcase';
 export { default as ReferralBlock } from './ReferralBlock/ReferralBlock';
-export { default as Refer } from './Refer/Refer';
+export { default as RefereeInput } from './RefereeInput/RefereeInput';
 export { default as Testimonials } from './Testimonials/Testimonials';
