--- src/components/views/Invite/Invite.scss
@@ -1,22 +1,0 @@
-@import '../../../styles/main';
-
-.react-cookie-banner {
-  text-align: left;
-}
-
-.cookie-message {
-  @include small;
-  @include flex;
-  @include padding(left, 2);
-  color: $white;
-}
-
-.cookie-link {
-  color: #b4b4b4;
-  letter-spacing: normal;
-  text-decoration: none;
-
-  &:hover {
-    color: $white;
-  }
-}
