+++ src/components/views/Referee/Referee.scss
@@ -1,0 +1,5 @@
+@import '../../../styles/main';
+
+.refereeContainer {
+  //
+}
