+++ src/components/views/index.js
@@ -1,6 +1,6 @@
 // Elements definition (ASC order)
 export { default as About } from './About/About';
-export { default as Invite } from './Invite/Invite';
 export { default as Legal } from './Legal/Legal';
 export { default as Privacy } from './Privacy/Privacy';
+export { default as Referee } from './Referee/Referee';
 export { default as Referral } from './Referral/Referral';
