+++ src/index.js
@@ -13,7 +13,7 @@ import About from './components/views/About/About';
 import Privacy from './components/views/Privacy/Privacy';
 import Legal from './components/views/Legal/Legal';
 import Home from './components/views/Home/Home';
-import Invite from './components/views/Invite/Invite';
+import Referee from './components/views/Referee/Referee';
 import Referral from './components/views/Referral/Referral';
 import AcceptReferral from './components/views/Referral/AcceptReferral';
 
@@ -30,7 +30,7 @@ ReactDOM.render(
         <Route path="/home" exact component={Home} />
         <Route path="/about" exact component={About} />
         <Route path="/privacy" exact component={Privacy} />
-        <Route path="/invite" exact component={Invite} />
+        <Route path="/referee" exact component={Referee} />
         <Route path="/legal" exact component={Legal} />
         <Route path="/statistics/:referralLink" exact component={Referral} />
         <Route
